requirements.txt
requests>=2.31.0
python-dotenv>=0.21.0
sentence-transformers>=2.2.0
chromadb>=0.3.21
qdrant-client>=1.9.0
PyPDF2>=3.0.0
numpy>=1.24.0
beautifulsoup4>=4.11.0
pdfplumber>=0.9.0
PyMuPDF>=1.23.0
cloudscraper>=1.2.71
google-generativeai
streamlit
pytest>=7.0.0